refactor(Application): main 로직을 calculatorService로 분리
- 여러 역할의 로직이 한번에 포함되서 이해하기 어려워서 로직을 분리하기 위해서
