+++ src/main/kotlin/calculator/Application.kt
@@ -7,13 +7,12 @@ fun main() {
     val calculator = Calculator()
     val numberTokenizer = NumberTokenizer()
 
-    guideOutput.guideDelimiters()
-    guideOutput.guideCustomDelimiters()
-    guideOutput.guideAdditionInput()
-    val input = userInput.getString()
-    val numbers = numberTokenizer.tokenize(input)
-    val customDelimiters = numberTokenizer.extractCustomDelimiters(input)
-    val sum = calculator.sum(numbers)
-    resultOutput.outputCustomDelimiters(customDelimiters)
-    resultOutput.outputAddition(sum)
+    val calculatorService = CalculatorService(
+        userInput = userInput,
+        guideOutput = guideOutput,
+        resultOutput = resultOutput,
+        calculator = calculator,
+        numberTokenizer = numberTokenizer
+    )
+    calculatorService.execute()
 }
